refactor(api): move hono endpoints into api package
