+++ packages/api/package.json
@@ -5,6 +5,7 @@
   "exports": {
     ".": "./src/index.ts",
     "./context": "./src/context.ts",
+    "./http": "./src/http/index.ts",
     "./r2": "./src/r2.ts",
     "./routers": "./src/routers/index.ts"
   },
@@ -15,10 +16,12 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.1068.0",
+    "@paralleldrive/cuid2": "^3.3.0",
     "@prive-admin-tanstack/auth": "workspace:*",
     "@prive-admin-tanstack/db": "workspace:*",
     "@prive-admin-tanstack/env": "workspace:*",
     "@trpc/server": "catalog:",
+    "archiver": "^8.0.0",
     "csv-parse": "^6.2.1",
     "drizzle-orm": "^0.45.2",
     "hono": "catalog:",
@@ -26,6 +29,7 @@
   },
   "devDependencies": {
     "@prive-admin-tanstack/config": "workspace:*",
+    "@types/archiver": "^6.0.3",
     "@types/node": "catalog:",
     "typescript": "catalog:"
   }
